Refactor to use integer based demand.

+++ src/main/java/com/bitlimit/bits/persistence/model/DemandLevel.java
@@ -20,9 +20,9 @@ public class DemandLevel extends Model
 	 *
 	 */
 
-	public Float getDemand()
+	public Integer getDemand()
 	{
-		return this.getFloat("demand");
+		return this.getInteger("demand");
 	}
 
 	/*
@@ -31,9 +31,9 @@ public class DemandLevel extends Model
 	 *
 	 */
 
-	public void setDemand(Float demand)
+	public void setDemand(Integer demand)
 	{
-		this.setFloat("demand", demand);
+		this.setInteger("demand", demand);
 	}
 
 	/*
@@ -47,6 +47,8 @@ public class DemandLevel extends Model
 		/* Discussion: this recursive function was a first attempt.
 		 * It will be revised to use a positive incrementing count which will then be fed into an arctan based function.
 		 */
+
+
 	}
